ridding beekit of the ui (components)
