+++ BeeSwift/GoalExtensions.swift
@@ -1,0 +1,7 @@
+// Part of BeeSwift. Copyright Beeminder
+
+import BeeKit
+
+extension Goal {
+  public var countdownColor: UIColor { return UIColor.Beeminder.SafetyBuffer.color(for: self.colorkey) }
+}
